DPL: refactor DataHeader to not include boost
This separates Stack in its own header so that boost is
not exposed when DataHeader is included. Apart from the discussion
of whether we should include boost in our ROOT_INCLUDE_PATH, this
is in any case a good refactoring which separates the data header
description from their actual manipulation.
