+++ Framework/Core/include/Framework/DataAllocator.h
@@ -412,21 +412,7 @@ public:
                                            o2::header::SerializationMethod serializationMethod, //
                                            size_t payloadSize);                                 //
 
-  Output getOutputByBind(OutputRef&& ref)
-  {
-    if (ref.label.empty()) {
-      throw std::runtime_error("Invalid (empty) OutputRef provided.");
-    }
-    for (size_t ri = 0, re = mAllowedOutputRoutes.size(); ri != re; ++ri) {
-      if (mAllowedOutputRoutes[ri].matcher.binding.value == ref.label) {
-        auto spec = mAllowedOutputRoutes[ri].matcher;
-        return Output{ spec.origin, spec.description, ref.subSpec, spec.lifetime, std::move(ref.headerStack) };
-      }
-    }
-    throw std::runtime_error("Unable to find OutputSpec with label " + ref.label);
-    assert(false);
-  }
-
+  Output getOutputByBind(OutputRef&& ref);
   void addPartToContext(FairMQMessagePtr&& payload,
                         const Output &spec,
                         o2::header::SerializationMethod serializationMethod);
