+++ Framework/Core/src/DataSamplingConditionNConsecutive.cxx
@@ -16,6 +16,7 @@
 #include "Framework/DataSamplingCondition.h"
 #include "Framework/DataSamplingConditionFactory.h"
 #include "Framework/DataProcessingHeader.h"
+#include <fairlogger/Logger.h>
 
 namespace o2
 {
