+++ Framework/Core/test/test_DataSamplingCondition.cxx
@@ -18,6 +18,7 @@
 #include "Framework/DataRef.h"
 #include "Framework/DataProcessingHeader.h"
 #include "Headers/DataHeader.h"
+#include "Headers/Stack.h"
 
 using namespace o2::framework;
 using namespace o2::header;
